Store the last TaskId in the RTC registers
We have an awkward problem where if some task is looping forever
we are stuck and can't access anything except via SWD. There
are some situatiosn (read racks) where we can't get dongle access.
We already have a watchdog which is used for reset.

The RTC block has a small set of backup registers that will not
be reset so long as we don't lose power to VDD (i.e. don't ignition
cycle). This makes our scheme:

- Setup our RTC block and use registers 0 and 1 for our purposes
- On context switch store the task ID into Register 0
- On bootup, copy the contents of Register 0 to Register 1
- In a separate task, set a timer to turn the watchdog off/on
- Log the value of Register 1 to see what the last task ID that
  was running was