Remove unused `#[used]` from things that are used (#2167)
To [paraphrase Donald Rumsfeld][1]: as we know, there are used useds;
there are things we use. We also know there are used unuseds, that is to
say there are some things we use but the compiler doesn't know we use,
so `#[used]` so that the thing we use appears used. . But there are also
unused useds; the ones we use but put `#[used]` on so the compiler
thinks we use them, but the `#[used]` is unused because we already use
it.

Presently, the `ringbuf!`, `counters!`, and `counted_ringbuf!` macros
all place a `#[used]` attribute on the statics that they generate. This
is intended to prevent the compiler from eliminating writes to these
statics when they are read only by humility. However, in the case of
both `ringbuf` and `counters`, the `#[used]` attribute is not actually
necessary: when writing to a ringbuf, we read the last value written, to
check if they are equal, and writes to a counter use `fetch_add`, which
is a load as well as a store. So, we don't need the `#[used]` attribute
in either case.

In recent nightly toolchains, the [behavior of `#[used]` on targets
which produce ELF binaries has changed][2]. Now, it also tells the
_linker_ that the value is used, as well as the compiler's own dead-code
elimination. This is unfortunate, as it means that ringbufs and counters
are never optimized out at link time, even when they are not actually
*written* to in a particular task. For example, consider the
`drv_i2c_devices` crate. This crate contains modules with drivers for a
whole bunch of different devices. Lots of tasks depend on
`drv_i2c_devices`, and use the drivers for one or two of those devices.
Most of the drivers have their own debugging ringbufs. When updating to
a nightly toolchain after the `#[used]` change, the ringbufs for *all*
the drivers in the crate get linked into *every* task that depends on
`drv_i2c_devices`, regardless of which drivers the task actually, well,
uses.

This commit fixes that by getting rid of the `#[used]` attributes. We
don't actually need them. I've flashed my Gimletlet with this change,
and all the ringbufs and counters are still there even without `#[used]`
on anything.

[1]: https://en.wikipedia.org/wiki/There_are_unknown_unknowns
[2]: https://github.com/oxidecomputer/hubris/issues/2165#issuecomment-3098126432