Make more ringbufs zero-initialized, saving flash
To simplify its implementation, ringbuf currently requires every ringbuf
to be declared with an initialization element. It creates a static array
containing that element.

If that element is represented as zero, that goes in BSS and is
zero-initialized with the rest of BSS.

If that element is represented as anything other than zero, the compiler
creates a full-sized "initialization image" copy of the ringbuf _in
flash_ that it can then memcpy into RAM on startup.

Using the default Rust enum representation, the easiest way to get a
ringbuf initialized to non-zero bytes is to put its initialization
sentinel value (e.g. None) as the last variant in the enum.
Unfortunately, this is super common.

This commit moves all the ones I could find, which should reduce flash
usage in a bunch of cases. A survey of examples in the Cosmo build
(note: many of these savings are not fully realized due to pow2
alignment restrictions):

- Cosmo Seq: -6328
- Spartan7 Loader: -2712
- I2C: 2488 bytes
- Every SPI driver: -1068 (so, -2136 total)
- Hiffy: -1564
- Sprot: -1076
- Update: -1036
- Validate: -1036
- Power: -559
- Thermal: -212
- Dump agent: -12 (...?)