`dump-agent` doesn't write ringbuf with `no-rot` (#2167)
Turns out the [compiler is actually right][1] that this enum is not used
when `dump-agent` is compiled with the `no-rot` feature flag, as we
don't write to the ringbuf at all when that feature is set. The
`#[used]` attribute was suppressing the unused type warning. Hahaha.
Thus, I just made the whole thing `#[cfg(not(feature = "no-rot"))]`.
Great.