Dead codes returned by `kipc::restart_task` are of the old generation (#2136)
When a task is restarted by `jefe`, it calls `kipc::restart_task`.

Within the kernel, any tasks blocked on the target task receive a dead
code.

Dead codes typically include the _new_ task generation, so that callers
can update their task ID. However, in this case, we return the older
generation. This adds an extra round-trip to idempotent calls:

- Task A calls into task B with generation 0
- Task B dies (RIP)
- `jefe` restarts the task, sending a dead code with generation 0 to
task A (**this is the bug**)
- Task A retries with generation 0 again
- Due to generation mismatch, the call fails, and Task A gets a dead
code with generation 1
- Task A retries with generation 1
- The call finally succeeds

This PR ensures that the dead code is from the new generation.

It also renames `kipc::restart_task` to `reinit_task`, to match the docs.

Finally, it adds a unit test for this behavior, which required tweaking
the test supervisor to optionally restart tasks.