stm32xx-i2c: remove I2cControl function table (#2121)
It turns out that, in practice, we only ever provide a single function
for each of `wfi` and `enable` here. This makes the table itself an
unnecessary level of indirection, which has some deleterious effects:

- It opens the possibility for _different_ I2cControl structs containing
_different_ function pointers to be passed to each call, complicating
analysis. (Searching for this phenomenon was what led me to notice that
we only ever use one implementation.)

- It makes it a lot harder to recognize use of syscalls (or misuse of
syscalls) in the I2C layer.

- It routes important parts of the I2C stack through indirect function
calls, defeating stack analysis.

- It adds a bunch of code that isn't really necessary, as it turns out.

This commit removes the dispatch table, inlining the existing functions
into their callsites (and introducing a utility function for the wfi
behavior).

---------

Co-authored-by: Matt Keeter <matt@oxide.computer>