Add delays to `jefe` restarts (#2129)
In the past, we have seen rapidly-crashing tasks perform a
denial-of-service attack on lower-priority tasks (#616, see also #1735).

This PR adds a `const MIN_RUN_TIME` to the `jefe` supervisor. If a task
crashes before that amount of time has elapsed, then its restart is
delayed to meet that minimum. I picked 50 ms arbitrarily, which limits a
task to 20 restarts per second.

Tracking task start times adds 16 bytes of additional state per task,
requiring more RAM for the supervisor (and in some cases bumping it to
the next size threshold).