Use refdes-based component IDs in control-plane-agent (#2199)
In the `gateway-sp-comms` protocol, hardware devices known to the SP are
uniquely identified by _component IDs_. These IDs are generated by the
SP and are given to the control plane when reporting inventory. The
control plane may then request details about a particular hardware
component using the ID received from the inventory.

Presently, these IDs use what I've been referring to as the "dev-NN
scheme". In this scheme, each component is identified by a string
consisting of `"dev-` followed by a two digit number, which is generated
by the `control-plane-agent` task from that device's position in an
array of devices generated at build time. While @jgallagher will claim
that this is just a bad design driven entirely by expedience, it does
have some redeeming qualities: the SP can look up devices very
efficiently by parsing the `NN` part as an integer and using it to index
the array, and all the identifiers are a fixed, short number of bytes.
However, it also has some disadvantages:

- The identifiers are not semantically meaningful
- A given device's identifier may change across firmware versions if new
devices are added to the `app.toml` or the ordering of devices changes
- The identifiers are known only to the `control-plane-agent` task, and
cannot easily be accessed by other tasks.

That last point is particularly problematic for the fault management
subsystem: we would like a variety of tasks to be able to generate
ereports relating to particular hardware components. When producing such
an ereport, we would like to be able to include an identifier for that
device that the control plane may then use to request additional
information about that device from the SP. These identifiers should also
be able to be used by the control plane to look up that device in its
own cached understanding of the SP's inventory, and for querying metrics
collected from the SP about that component. Since the `dev-nn` component
IDs are known only to the `control-plane-agent` task, other Hubris tasks
cannot easily include them in ereports. This is discussed at length in
#2197.

In that issue, we [decided][1] to solve this problem by changing the
`control-plane-agent` task from identifying components using the
`dev-NN` scheme to a new scheme based on reference designators (refdes).
Since a device's refdes is set in the `app.toml` for the board, it can
be accessed by any number of tasks if we include it in generated code
for the I<sup>2</sup>C (and, eventually, SPI) APIs. We started on this
in #2189. Using refdes as the identifiers for hardware components has a
number of other advantages:

- In `host-sp-comms`, we report inventory to the host OS using refdes as
the unique ID for devices, so this makes the MGS API consistent with the
IPCC API.
- Refdes is a fixed property of the _board_, rather than a property of
the particular layout of items in a TOML file. This means that it should
not change across Hubris versions (barring a bug or incorrect refdes in
an app config).
- The refdes indicates semantically meaningful information; it can be
used to find a given component in the board's schematic.
- Refdes are also pretty short strings

This PR changes `control-plane-agent` to use the device's refdes as the
component ID, rather than the `dev-nn` identifier. When looking up a
device, we use an array mapping refdes strings to indices in the devices
array, which is generated at build time. We perform a binary search over
this array to find the index and then access the device by index.

Certain devices (typically, those with updateable firmware) use
hard-coded component ID strings. On Gimlets, for example, these are
`sp`, `sp3-host-cpu`, `host-boot-flash`, and `system-led`. I left these
alone and didn't mess with them.

In a handful of cases, multiple devices have the same refdes in
`app.toml`. Typically, this is when a removable subassembly is on the
"other side" of a connector --- for example, a sharkfin has a VPD
EEPROM, a MAX5970 hotswap controller, and the U.2 device's NVMe Basic
Management Command, and these all have refdes like `J206`, for the CEM
slot that the sharkfin occupies. In this case, the `host-sp-comms` task
uses a scheme like `{connector_refdes}/{refdes_on_subassembly}`, so (for
example) sharkfin A's VPD EEPROM is `J206/U7` and its HSC is `J206/U8`.
Currently, this is all just hardcoded in `host-sp-comms`. In order to
use this scheme in `control-plane-agent` as well, I've changed the toml
format to accept refdes either as a string _or_ an array of strings,
which are joined together at codegen-time with `/` when forming the
device IDs, or `_`s when generating Rust identifiers. There were some
preexisting cases of refdes strings like `Jxxx_Uxx` in some app.tomls;
I've changed these to use the new format.

Finally, there were a handful of boards which lacked refdes for some or
all of their devices. After this change, these boards no longer build,
so I've gone and updated them.

Fixes #2197

[1]:
https://github.com/oxidecomputer/hubris/issues/2197#issuecomment-3192600805

---------

Co-authored-by: Matt Keeter <matt@oxide.computer>