i2c: make component IDs non-optional (#2203)
This commit changes `build-i2c` and `drv-i2c-api` so that when
refdes-based component IDs are generated, they are no longer optional.
This way, we can rely on them being present. This required a change in
`drv-sensor-api` to actually enable component IDs there.

Also, I renamed the config setting from `include_refdes` to
`component_ids` for consistency.