Swap FMC base address
The default attributes for the FMC region are normal and executable.
This allows speculative accesses. This doesn't apply to most hubris
tasks but it does apply to the kernel. Swap the FMC banks to use
a different default memory map which marks it as device/execute never
to avoid speculation. As part of this, fix up the FMC demo server
to translate addresses to the new region.