cpu-seq: Vcore PMBus alert ereports (#2184)
This branch implements ereports for PMBus input fault and warning alerts
in the Cosmo and Gimlet sequencers. There's kind of a lot going on here.
Essentially, I've hacked up @bcantrill's `drv_gimlet_seq_server::vcore`
module to produce an ereport when we detect a PMBus input alert. These
ereports contain the value of PMBus status registers read from the VRM,
as well as whether or not the VRM claims `POWER_GOOD` has been
deasserted, and identity information of the device ID and power rail. On
Cosmo, I've added similarish code on top of the sequencer interrupt
handling code @labbott added in #2181. It's a bit more complex on Cosmo
as there are two RAA22960A VRMs driving the `VDDCR_CPU0` and
`VDDCR_CPU1` rails, rather than a single regulator as in SP3's single
`VDD_VCORE` rail.

This branch closes #2188, which has been merged into this PR.

Closes #2141