Speed up FPGA loading on Cosmo / Grapefruit (#2213)
On Cosmo and Grapefruit, it takes ~10 seconds to load the sequencer
FPGA, which is a very human-noticeable delay.

According to https://github.com/oxidecomputer/hubris/issues/2088 , 4-bit
QSPI works on Cosmo at 66 MHz, so this PR adds a new `fast-qspi` feature
to use those settings. The feature is enabled for Cosmo and Grapefruit.
It is not enabled for Sidecar because that hasn't been validated, and
Sidecar load times have been less annoying.

In addition, the PR bumps our FPGA SPI clock divider from `DIV64`
(1.5625 MHz) to `DIV8` (12.5 MHz). This is well below the FPGA's limit
of 100 MHz, but any faster and we hit
https://github.com/oxidecomputer/hubris/issues/2133

Together, this bring the FPGA load time down to about 4 seconds.

I tested this on `cosmo-hubris-sp` and confirmed that the host boots up
to `bldb`, which is a strong indicator that everything is working.