packrat: Add helper API for `serde` ereports (#2225)
In practice, so far, most ereport payloads have been defined as structs
implementing `serde::Serialize` and encoded using the `minicbor-serde`
adapter crate. This requires a bit of boilerplate code, so it seems
worthwhile to provide a method that just does that. This commit adds one
in `task-packrat-api`. It's feature-flagged so we can avoid compiling
`minicbor`, `minicbor-serde`, and `serde` in tasks that don't submit
ereports using this.