Disable SWD pins when debugger is connected (#2228)
Fixes #2222

This PR adds state to the RoT's SWD task, tracking whether it is
"connected" (pins configured for SPI) or "disconnected" (pins configured
as GPIOs). When the `JTAG_DETECT` line fires, we switch to disconnected;
when we attempt to communicate with the SP, we try to reconnect.

There are two new generated functions: `switch_io_connected` and
`switch_io_disconnected`. In addition, I extracted the repeated "build a
helper function to set pins" logic in the build script.

I tested this on a grapefruit by resetting the SP with / without
debugger attached, and confirmed that measurements were made when the
debugger was detached. In addition, I could see in the RoT ringbufs that
it detected the presence of the debugger and switched IO pins to inputs.

Other small cleanups:
- Line wrapping comments
- Remove the unused `db_reset_sp` function