stm32xx-sys: fix glitch on initial pin configuration
I wound up here while hunting #1824, which I had assumed to be
I2C-specific. It isn't! It just isn't visible on most other interfaces
for subtle reasons.

The GPIO setup code was writing registers in the following order:

1. MODER (selecting from: input, output, alternate, analog)
2. OTYPER (selecting push-pull vs open-drain)
3. OSPEEDR (selecting slew rate)
4. PUPDR (enabling optional pull resistors)
5. AFRx (choosing an alternate function

At reset, the vast majority of GPIOs are in analog mode. This write
order meant that, when switching any pin to Alternate mode after reset,
we would briefly expose AF0 before switching to the correct Alternate
Function mux value. AF0 is not actually implemented on most pins, and in
practice seems to be equivalent to a constant 0 output.

This meant that in cases where a pin has an external pullup resistor,
the pin would show a very brief negative pulse before releasing. This
was most obvious on I2C, where external pullup resistors are standard,
and where a random negative glitch on a pin is Bad, Actually (per the
standard).

This commit reorders the register writes and more clearly enumerates the
conditions it's trying to uphold. When switching between two arbitrary
peripherals, you need to go through something like Analog mode to avoid
a visible glitch -- but afaict we never do this. When switching between
a peripheral and Analog mode for tristating I2C pins, the new order
ensures glitch-free operation, which I've verified on my logic analyzer.

Fixes #1824.