stm32xx-gpio-common: fix another glitch
My original fix only made a difference when switching _into_ Alternate
mode. In the relatively common case of switching _out of_ Alternate into
e.g. Analog or Input (to tristate the pin), code usually uses a utility
wrapper that passes AF0 to gpio_configure_raw... and even with the fix
in place, we'd update AFRx to AF0 before changing MODER.

Which is the right thing to do if you're going _to_ Alternate, but not
vice versa.

So, we got brief glitches to AF0 (which in practice on most pins means
pulled to ground) when attempting to tristate pins.

It's possible that we need to revisit this API, which is arguably overly
clever. But for now, this commit changes the logic so we don't update
AFRx unless relevant.

Fixes #1824 (again).