ADM1273 support (#2250)
This change is largely a renaming of ADM1272 to ADM127x since the
ADM1273 is a drop-in replacement and has the same behavior. The only
thing of substance that was modified was how we Validate this device.
The current behavior is the read the MFG_MODEL register and expect it to
be `ADM1272-2A`. I think this is a bit fragile as the die revision is
tacked on there (`-2A`) and could change without us knowing (the
orderable PN here is actually still `ADM1272-1A` even though
`ADM1272-2A` is being delivered). Then there is the additional
complication that with the ADM1273 we would see something like
`ADM1273-$A`. Thus, I've changed the behavior of the `adm127x::validate`
no longer use the common `pmbus_validate` but to use its own version
which will look at the first 7 bytes of the MFG_MODEL and make sure it is
either `ADM1272` or `ADM1273`.

fixes #2247