Support `MPN1` barcodes in `oxide-barcode`/`host-sp-comms` (#2229)
As described in #2209, the `MPN1` serial scheme is used for FRUs which
have manufacturer-issued serial numbers. This format is described in
[§7.2 RFD 308]. This format will be used in upcoming batches of Gimlet
and Cosmo fan tray assemblies (see hardware-rack#368), so Hubris must be
able to support the MPN1 scheme.

This commit adds support for parsing MPN1 serial numbers in the
`oxide-barcode` crate, and adds new IPCC messages to `host-sp-messages`
to support MPN1 serial numbers in fan trays. While we use a structured
representation for Oxide OXV1/OXV2 barcodes, where individual components
of the barcode (the serial, revision, and part number) are stored in
separate fields, we choose to represent MPN1 serials as a single string.
Unlike the OXV1/OXV2 barcodes, where the part number and serial number
are always 11 characters long, and the revision is always 3 characters,
the primary length limitation of the MPN1 format is that the *entire
barcode* will not exceed 128 characters.[^1] Therefore, parsing out
these individual fields would require a 114-byte buffer for each
component, even though none of them will be that long. Instead, we've
chosen not to separate out the fields. See [here] for details.

Since MPN1 serials cannot be represented by the existing `Identity`
struct in `host-sp-messages`, which is used for the current
`InventoryData::FanIdentity` IPCC message, we must introduce a new
`FanIdentityV2` mesage to represent fan trays that are serialized with
either format. The SP will now always send both fan tray inventory
messages, but if MPN1 serialized fans are encountered, their identities
will be left empty in the V1 message. This is to ensure that we don't
send an incomplete identity that may collide with other serial numbers
(since MPN1 serials are only guaranteed to be unique in the context of
the manufacturer code, part number, and revision). Host software will be
updated to prefer the `FanIdentityV2` message when both are present, as
it will always contain all the serials, but we will still report some
fan tray identity information in a format that previous host software
versions will understand.

Fixes #2209

[§7.2 RFD 308]: https://rfd.shared.oxide.computer/rfd/308#fmt-mpn
[here]:
  https://github.com/oxidecomputer/hubris/issues/2209#issuecomment-3320601513
[^1]: Additionally, the manufacturer code is always three characters, as
      it's the one part of the format assigned by Oxide.